requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
asyncpg
supabase
pydantic>=2.0
pydantic-settings
python-multipart 
jinja2
itsdangerous
extract-msg
openai